+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 * Allow `LoadResource` inject into Set of strings #470
 * Param binding - it shall not try to get provider for simple types #449
 * Support plugin JAX-RS #448
+* Add method annotation to direct framework use `JSON.toJSONStringWithDateFormat()` #281
 * Trace handler call #238
 * Make it able to specify the named ports in routes configuration #183